fix:update on candidate and recrutier dashboard - fix issue with broken links - fix naming convention for components - refactor code
