+++ components/Dashboard/candidate/applied/index.tsx
@@ -1,0 +1,12 @@
+import AdminLayout from "@/layouts/adminLayout";
+
+export default function Applied(): JSX.Element {
+    return (
+        <AdminLayout>
+            hello applied
+        </AdminLayout>
+
+    )
+}
+
+
