+++ components/Dashboard/candidate/messages/index.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+import AdminLayout from "layouts/adminLayout";
+
+export default function Messages(): JSX.Element {
+    return <AdminLayout>Candidate Messages page</AdminLayout>;
+};
+
