+++ components/Dashboard/candidate/settings/notifications.tsx
