+++ components/Dashboard/recruiter/company/index.tsx
@@ -1,0 +1,17 @@
+import React from "react";
+import RecruiterProfileLayout from "@/layouts/recruiterProfileLayout";
+
+
+
+
+const Company = ():JSX.Element => {
+    return (
+        <RecruiterProfileLayout>
+            <p>Recruiter company profile</p>
+
+        </RecruiterProfileLayout>
+
+    )
+}
+
+export default Company
