+++ components/Dashboard/recruiter/forum/index.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+import AdminLayout from "layouts/adminLayout";
+
+const Forum = ():JSX.Element => {
+    return <AdminLayout>Recrutier Forum page</AdminLayout>;
+};
+
+export default Forum;
