+++ components/Dashboard/recruiter/jobs/index.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+import AdminLayout from "layouts/adminLayout";
+
+const Jobs = ():JSX.Element => {
+    return <AdminLayout>Recrutier Jobs page</AdminLayout>;
+};
+
+export default Jobs;
