+++ components/Dashboard/recruiter/messages/index.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+import AdminLayout from "layouts/adminLayout";
+
+const Messages = (): JSX.Element => {
+    return <AdminLayout>Recrutier Messages page</AdminLayout>;
+};
+
+export default Messages;
