--- components/dashboard/candidate/applied/index.tsx
@@ -1,14 +1,0 @@
-import AdminLayout from "@/layouts/adminLayout";
-import { NextPage } from "next";
-
-const Applied: NextPage = () => {
-    return (
-        <AdminLayout>
-            hello applied
-        </AdminLayout>
-
-    )
-}
-
-
-export default Applied
