--- components/dashboard/candidate/messages/index.tsx
@@ -1,9 +1,0 @@
-import React from "react";
-import { NextPage } from "next";
-import AdminLayout from "layouts/adminLayout";
-
-const Messages: NextPage = () => {
-    return <AdminLayout>Candidate Messages page</AdminLayout>;
-};
-
-export default Messages;
