--- components/dashboard/recruiter/company/index.tsx
@@ -1,18 +1,0 @@
-import React from "react";
-import { NextPage } from "next";
-import RecruiterProfileLayout from "@/layouts/recruiterProfileLayout";
-
-
-
-
-const Company: NextPage = () => {
-    return (
-        <RecruiterProfileLayout>
-            <p>Recruiter company profile</p>
-
-        </RecruiterProfileLayout>
-
-    )
-}
-
-export default Company
