--- components/dashboard/recruiter/forum/index.tsx
@@ -1,9 +1,0 @@
-import React from "react";
-import { NextPage } from "next";
-import AdminLayout from "layouts/adminLayout";
-
-const Forum: NextPage = () => {
-    return <AdminLayout>Recrutier Forum page</AdminLayout>;
-};
-
-export default Forum;
