--- components/dashboard/recruiter/jobs/index.tsx
@@ -1,9 +1,0 @@
-import React from "react";
-import { NextPage } from "next";
-import AdminLayout from "layouts/adminLayout";
-
-const Jobs: NextPage = () => {
-    return <AdminLayout>Recrutier Jobs page</AdminLayout>;
-};
-
-export default Jobs;
