--- components/dashboard/recruiter/messages/index.tsx
@@ -1,9 +1,0 @@
-import React from "react";
-import { NextPage } from "next";
-import AdminLayout from "layouts/adminLayout";
-
-const Messages: NextPage = () => {
-    return <AdminLayout>Recrutier Messages page</AdminLayout>;
-};
-
-export default Messages;
