+++ pages/dashboard/company/index.tsx
@@ -1,0 +1,15 @@
+import React from "react";
+import { NextPage } from "next";
+import RecruiterCompany from '@/components/Dashboard/recruiter/company/index';
+
+
+const Company: NextPage = () => {
+    return (
+        <>
+            <RecruiterCompany />
+        </>
+    )
+}
+
+
+export default Company
