+++ pages/dashboard/forum/mentors.tsx
@@ -1,0 +1,14 @@
+import React from "react";
+import { NextPage } from "next";
+import CandidateMentors from "@/components/Dashboard/candidate/forum/mentors";
+
+const Mentors: NextPage = () => {
+    return (
+        <>
+            <CandidateMentors />
+        </>
+    )
+}
+
+
+export default Mentors
