+++ pages/dashboard/forum/splash.tsx
@@ -1,0 +1,14 @@
+import React from "react";
+import { NextPage } from "next";
+import CandidateSplash from "@/components/Dashboard/candidate/forum/splash";
+
+const Splash: NextPage = () => {
+    return (
+        <>
+            <CandidateSplash />
+        </>
+    )
+}
+
+
+export default Splash
