+++ pages/dashboard/index.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { NextPage } from "next";
-import Candidate from "@/components/dashboard/candidate";
-import Recruiter from "@/components/dashboard/recruiter";
+import Candidate from "@/components/Dashboard/candidate";
+// import Recruiter from "@/components/Dashboard/recruiter";
 
 const Dashbaord: NextPage = () => {
     return (
