+++ pages/dashboard/jobs/index.tsx
@@ -1,0 +1,16 @@
+import React from "react";
+import { NextPage } from "next";
+import CandidateJobs from "@/components/Dashboard/candidate/jobs";
+import RecruiterJobs from "@/components/Dashboard/recruiter/jobs";
+
+const Jobs: NextPage = () => {
+    return (
+        <>
+            <CandidateJobs />
+            {/* <RecruiterJobs/> */}
+        </>
+    )
+}
+
+
+export default Jobs
