+++ pages/dashboard/settings/index.tsx
@@ -1,0 +1,18 @@
+import React from "react";
+import { NextPage } from "next";
+import CandidateSettings from "@/components/Dashboard/candidate/settings/index";
+import RecruiterSettings from "@/components/Dashboard/recruiter/settings/index";
+
+
+
+const Settings: NextPage = () => {
+    return (
+        <>
+            <CandidateSettings />
+            {/* <RecruiterSettings/> */}
+        </>
+    )
+}
+
+
+export default Settings
