+++ pages/dashboard/settings/notifications.tsx
@@ -1,0 +1,18 @@
+import React from "react";
+import { NextPage } from "next";
+import CandidateNotification from "@/components/Dashboard/candidate/settings/notifications";
+import RecruiterNotification from "@/components/Dashboard/recruiter/settings/notifications";
+
+
+
+const Notification: NextPage = () => {
+    return (
+        <>
+            <CandidateNotification />
+            {/* <RecruiterNotification/> */}
+        </>
+    )
+}
+
+
+export default Notification
